<b>Fix 501 bug, remove empty query and post</b>
<a href="http://wiki.tcl.tk/_/revision?N=27377&V=21">Source</a>, <a href="http://wiki.tcl.tk/_/revision?N=27383&V=29#pagetoc41bb1c27">announcement</a>.

Oops.  The 501 code didn't work.  I forgot to quote the value of the header key, resulting in an extra argument to [dict create].  Also, I removed the query, rawquery, post, and rawpost keys from the request dictionary when no query or post was made.  This makes it easier to check if there was a query or a post.

  *  Fix 501 bug.
  *  Remove empty query, rawquery, post, and rawpost dicts.